Disable parser tests for categories count
When inserting article, MediaWiki core uses a dummy parser that does not
process link.  The parser tests having [[Category:test pages]] ends up
being broken because the dummy parser does not attach the created
article to a category (wikilins are skippd, the category is not added to
'categorylinks').

The dummy parser in MediaWiki is 65b732f29.

Until a workaround is available, disable two parser tests:

 WCategory->pageCount Test pages
 WCategory->subcatCount

Bug: T163733
Change-Id: I38da5a61d52dc6cd282997896688e39803caa895